Fix cast warning

+++ ManiVault/src/actions/WidgetAction.h
@@ -217,7 +217,7 @@ public: // Hierarchy queries
      */
     template<typename WidgetActionType = WidgetAction>
     std::uint32_t getNumberOfChildren(bool recursively = false) const {
-        return getChildren<WidgetActionType>(recursively).count();
+        return static_cast<std::uint32_t>(getChildren<WidgetActionType>(recursively).count());
     }
 
     /**
